requirements.txt
colorlog==6.7.0
homeassistant==2023.7.3
pip>=21.0,<23.2
ruff==0.0.284
pytest-homeassistant-custom-component>=0.13.50
pytest-asyncio==0.21.0
